pyproject.toml
[tool.poetry]
name = "tls_fragment"
version = "3.1.0"
description = "tls fragment"
readme = "README.md"
authors = ["Maoist2009 <maoist2009@example.com>"]
license = "MIT"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.11"
ahocorasick-python="^0.0.9"
requests="^2.32.3"
dnspython="^2.3.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
isort = "^5.12.0"
pytest = "^7.3.1"
tox = "^4.5.2"
mkdocs = "^1.4.3"
mkdocs-material = "^8.5.11"
pytest-pylint = "^0.19.0"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.pytest.ini_options]
testpaths = "tests"
python_files = "tests.py test_*.py *_tests.py"

[tool.pylint.design]
max-line-length = 120

[tool.poetry.plugins.console_scripts]
tls_fragment = "tls_fragment.cli:start_server"